pyproject.toml
[project]
name = "ozrv-2025-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "k",email = "iliakripa@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg (>=0.30.0,<0.31.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "aiohttp (>=3.11.16,<4.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "hf-xet (>=1.0.3,<2.0.0)",
    "greenlet (>=3.2.0,<4.0.0)",
    "psycopg (>=3.2.6,<4.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "uvicorn (>=0.34.1,<0.35.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "cryptography (>=44.0.2,<45.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
